requirements.txt
flask==2.0.1 
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SeaSurf==0.3.1
Flask-WTF==0.15.1
Flask-Admin==1.5.8
Flask-Security-Too==4.1.2
Flask-Babel
cssmin==0.2.0
jsmin==2.2.2
pyotp==2.6.0
email_validator==1.1.3
netifaces==0.11.0 
requests==2.26.0
bcrypt==3.1.4
pyqrcode==1.2.1
cryptography==35.0.0
bleach==4.1.0
gunicorn==20.1.0
